langchain/requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
arxiv
PyMuPDF
wikipedia
langchain-text-splitters
langchain-ollama
sentence_transformers
langchain-huggingface
faiss-cpu
langchain_chroma
langchain-classic
langchain_groq
uvicorn
fastapi
langserve
sse_starlette
langchain-google-genai
langgraph
torchvision
ddgs
psycopg2
validators
youtube-transcript-api
pytube
unstructured
numexpr
huggingface_hub
langchain-astradb
astrapy
datasets
tiktoken
pypdf2
cassandra-driver
gradio
langchain_nvidia_ai_endpoints
crewai
crewai_tools
pytubefix